refactor and update codebase
Refactor m8 into its own class. It has two types of macros, external and
internal. External macros are defined in a json config file, which is
parsed at runtime. Any program can be used as an external macro. The
internal macros are defined in ./src/macros.cc. The type of the macro is
stored in the macro struct, along with other info about the macro.

Define an interface to set m8's settings, such as config file location,
delimiters, and debug output.

If no output file is given as an argument, the default action is to
print to stdout.

The summary output is sent to stderr.

The default action when an invalid macro or regex is found is to replace
it with an empty string.